+++ src/main/java/codesquad/springcafe/web/repository/JdbcTemplateArticleRepository.java
@@ -49,7 +49,7 @@ public class JdbcTemplateArticleRepository implements ArticleRepository{
 
 
     @Override
-    public Optional<Article> findByIndex(int articleId) {
+    public Optional<Article> findByIndex(Long articleId) {
         String sql = "SELECT * FROM articles WHERE articleId = ?";
         List<Article> articles = jdbcTemplate.query(sql, new BeanPropertyRowMapper<>(Article.class), articleId);
         return articles.isEmpty() ? Optional.empty() : Optional.of(articles.get(0));
